% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liptak.R
\name{liptak}
\alias{liptak}
\title{Liptak's weighted z-test}
\usage{
liptak(x, y, alternative = c("greater", "less", "two.sided"))
}
\arguments{
\item{x}{first sample}

\item{y}{second sample}

\item{alternative}{comparison in alternative hypothesis (two-sided, less, or greater)}
}
\value{
P-value for the test
}
\description{
Calculates Liptak's weighted z-test p-value for partially matched samples.
}
\examples{
liptak(hetNA[,1],hetNA[,2],alternative="two.sided")

}
