% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looneyjones.R
\name{looneyjones}
\alias{looneyjones}
\title{Looney and Jones' corrected Z-test}
\usage{
looneyjones(x, y, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{first sample}

\item{y}{second sample}

\item{alternative}{comparison in alternative hypothesis (two-sided, less, or greater)}
}
\value{
Vector of length two whose first element is the test statistic and second element is the p-value
}
\description{
Calculates Looney and Jones' corrected Z-test test statistic and corresponding p-value for partially matched samples.
}
\examples{
looneyjones(hetNA[,1],hetNA[,2],alternative="two.sided")

}
